add reducer and action (shafa)
